* GMP size issues

+++ src/pl-arith.c
@@ -128,7 +128,7 @@ static int		registerFunction(ArithFunction f, int index);
 static int		getCharExpression(term_t t, Number r ARG_LD);
 static int		ar_sign_i(Number n1);
 static int		ar_add(Number n1, Number n2, Number r);
-static int		ar_add_ui(Number n, intptr_t val);
+static int		ar_add_ui(Number n, long val);
 static int		ar_minus(Number n1, Number n2, Number r);
 
 
@@ -878,7 +878,7 @@ promoteIntNumber(Number n)
 		*********************************/
 
 static int
-ar_add_ui(Number n, intptr_t add)
+ar_add_ui(Number n, long add)
 { switch(n->type)
   { case V_INTEGER:
     { int64_t r = n->value.i + add;
@@ -895,9 +895,9 @@ ar_add_ui(Number n, intptr_t add)
 #ifdef O_GMP
     case V_MPZ:
     { if ( add > 0 )
-	mpz_add_ui(n->value.mpz, n->value.mpz, add);
+	mpz_add_ui(n->value.mpz, n->value.mpz, (unsigned long)add);
       else
-	mpz_sub_ui(n->value.mpz, n->value.mpz, -add);
+	mpz_sub_ui(n->value.mpz, n->value.mpz, (unsigned long)-add);
 
       succeed;
     }
@@ -1054,7 +1054,7 @@ msb64(int64_t i)
 static int
 ar_shift(Number n1, Number n2, Number r, int dir) 
 { GET_LD
-  intptr_t shift;
+  long shift;
   const char *plop = (dir < 0 ? "<<" : ">>");
 
   if ( !toIntegerNumber(n1) ) 
@@ -1073,7 +1073,7 @@ ar_shift(Number n1, Number n2, Number r, int dir)
 	   n2->value.i > LONG_MAX )
 	return (int)outOfStack((Stack)&LD->stacks.global, STACK_OVERFLOW_RAISE);
       else
-	shift = (intptr_t)n2->value.i;
+	shift = (long)n2->value.i;
       break;
 #ifdef O_GMP
     case V_MPZ:
@@ -1233,13 +1233,13 @@ ar_pow(Number n1, Number n2, Number r)
 { 
 #ifdef O_GMP
   if ( intNumber(n1) && intNumber(n2) )
-  { uintptr_t exp;
+  { unsigned long exp;
 
     switch(n2->type)
     { case V_INTEGER:
 	if ( n2->value.i < 0 || n2->value.i > LONG_MAX )
 	  goto doreal;
-	exp = (uintptr_t)n2->value.i;
+	exp = (long)n2->value.i;
 	break;
       case V_MPZ:
 	if ( mpz_sgn(n2->value.mpz) < 0 ||
@@ -1258,7 +1258,7 @@ ar_pow(Number n1, Number n2, Number r)
     switch(n1->type)
     { case V_INTEGER:
 	if ( n1->value.i >= 0L && n1->value.i <= LONG_MAX )
-	{ mpz_ui_pow_ui(r->value.mpz, (intptr_t)n1->value.i, exp);
+	{ mpz_ui_pow_ui(r->value.mpz, (unsigned long)n1->value.i, exp);
 	  succeed;
 	} else
 	{ promoteToMPZNumber(n1);
